Factored input and dither structs into separate headers.
