+++ main.cpp
@@ -1,39 +1,5 @@
 /*
-
-Copyright 2009, 2011 Sebastian Gesemann. All rights reserved.
-
-Redistribution and use in source and binary forms, with or without modification, are
-permitted provided that the following conditions are met:
-
-   1. Redistributions of source code must retain the above copyright notice, this list of
-      conditions and the following disclaimer.
-
-   2. Redistributions in binary form must reproduce the above copyright notice, this list
-      of conditions and the following disclaimer in the documentation and/or other materials
-      provided with the distribution.
-
-THIS SOFTWARE IS PROVIDED BY SEBASTIAN GESEMANN ''AS IS'' AND ANY EXPRESS OR IMPLIED
-WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND
-FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SEBASTIAN GESEMANN OR
-CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
-ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
-NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
-ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
-The views and conclusions contained in the software and documentation are those of the
-authors and should not be interpreted as representing official policies, either expressed
-or implied, of Sebastian Gesemann.
-
- */
-/* ========================================
- *  NotJustAnotherDither - NotJustAnotherDither.h
- *  Copyright (c) 2016 airwindows, Airwindows uses the MIT license
- * ======================================== */
-
-/*
- Modifications and additions Copyright (c) 2023 clone206
+ Copyright (c) 2023 clone206
 
  This file is part of dsd2dxd
 
@@ -57,412 +23,31 @@ or implied, of Sebastian Gesemann.
 #include "AudioFile.h"
 #include "dsdin.hpp"
 #include "Output.hpp"
+#include "Dither.hpp"
+#include "Input.hpp"
 
 using namespace std;
 namespace fs = std::filesystem;
 
 #define DSD_64_RATE 2822400
 
+bool verboseMode = false;
+
+inline void verbose(string say, bool newLine = true)
+{
+    if (verboseMode)
+    {
+        cerr << say << (newLine ? "\n" : "");
+    }
+}
+
+inline bool lowercmp(char a, char b)
+{
+    return tolower(a) == b;
+}
+
 namespace
 {
-    bool verboseMode = false;
-
-    inline void verbose(string say, bool newLine = true)
-    {
-        if (verboseMode)
-        {
-            cerr << say << (newLine ? "\n" : "");
-        }
-    }
-
-    inline bool lowercmp(char a, char b)
-    {
-        return tolower(a) == b;
-    }
-
-    struct InputContext
-    {
-        int lsbitfirst;
-        bool interleaved;
-        bool stdIn;
-        int dsdRate;
-        string input;
-        fs::path filePath;
-
-        int dsdStride;
-        int dsdChanOffset;
-        int channelsNum;
-        int blockSize;
-        long audioLength;
-        off_t audioPos;
-        TagLib::PropertyMap props;
-
-        InputContext() {}
-
-        InputContext(string inFile, char fmt, char endianness, int inRate, int blockSizeIn, int channels)
-        {
-            if (lowercmp(endianness, 'l'))
-            {
-                lsbitfirst = 1;
-            }
-            else if (lowercmp(endianness, 'm'))
-            {
-                lsbitfirst = 0;
-            }
-            else
-            {
-                throw "No endianness detected!";
-            }
-
-            if (lowercmp(fmt, 'p'))
-            {
-                interleaved = false;
-            }
-            else if (lowercmp(fmt, 'i'))
-            {
-                interleaved = true;
-            }
-            else
-            {
-                throw "No fmt detected!";
-            }
-
-            if (inRate != 1 && inRate != 2)
-            {
-                throw "Unsupported DSD input rate.";
-            }
-
-            input = inFile; // "-" == stdin
-            channelsNum = channels;
-            dsdRate = inRate;
-            setBlockSize(blockSizeIn);
-            props = TagLib::PropertyMap();
-
-            if (input == "-")
-            {
-                stdIn = true;
-                audioLength = 0;
-                audioPos = 0;
-            }
-            else
-            {
-                stdIn = false;
-                filePath = fs::path(input);
-                verbose("Input file basename: ", false);
-                verbose(filePath.stem());
-                verbose("Parent path: ", false);
-                verbose(fs::absolute(filePath).parent_path());
-
-                FILE *inFile;
-                if ((inFile = fopen(input.c_str(), "rb")) != NULL)
-                {
-                    auto myDsd = dsd(inFile);
-                    audioPos = myDsd.audioPos;
-                    audioLength = myDsd.audioLength;
-                    channelsNum = myDsd.channelCount;
-                    dsdRate = myDsd.dsdRate;
-                    interleaved = myDsd.interleaved;
-                    lsbitfirst = myDsd.isLsb;
-
-                    if (myDsd.blockSize)
-                    {
-                        verbose("Setting block size from file");
-                        setBlockSize(myDsd.blockSize);
-                    }
-                    verbose("Audio length in bytes: ", false);
-                    verbose(std::to_string(audioLength));
-                }
-
-                TagLib::FileRef f(input.c_str());
-
-                if (!f.isNull() && f.tag())
-                {
-                    TagLib::Tag *tag = f.tag();
-                    verbose("Artist: ", false);
-                    verbose(tag->artist().to8Bit());
-                    props = f.properties();
-                }
-            }
-        }
-
-        void setBlockSize(int blockSizeIn)
-        {
-            blockSize = blockSizeIn;
-            dsdChanOffset = interleaved ? 1 : blockSize; // Default to one byte for interleaved
-            dsdStride = interleaved ? channelsNum : 1;
-        }
-    };
-
-    struct Dither
-    {
-        double noise_shaping_l; // Noise shape state L
-        double noise_shaping_r; // Noise shape state R
-        double byn_l[13];       // Benford Real number weights L
-        double byn_r[13];       // Benford Real number weights R
-        uint32_t fpd;           // Floating-point dither
-
-        char type;
-
-        inline void processSamp(double &sample, int chanNum);
-        inline void fpdither(double &sample);
-        inline void njad(double &sample, int chanNum);
-        inline void tpdf(double &sample);
-
-        Dither() {}
-
-        Dither(char ditherTypeOut)
-        {
-            type = tolower(ditherTypeOut);
-
-            if (type != 'n' && type != 't' && type != 'f' && type != 'x')
-            {
-                throw "Invalid dither type!";
-            }
-        }
-
-        void init()
-        {
-            if (type != 'x')
-            {
-                // Seed rng
-                srand(static_cast<unsigned>(time(0)));
-            }
-
-            if (type == 'n')
-            {
-                initOutputs();
-            }
-            else if (type == 'f')
-            {
-                initRand();
-            }
-        }
-
-        // Initialize outputs/dither state
-        void initOutputs()
-        {
-            // Weights based on Benford's law. Smaller leading digits more likely.
-            byn_l[0] = 1000;
-            byn_l[1] = 301;
-            byn_l[2] = 176;
-            byn_l[3] = 125;
-            byn_l[4] = 97;
-            byn_l[5] = 79;
-            byn_l[6] = 67;
-            byn_l[7] = 58;
-            byn_l[8] = 51;
-            byn_l[9] = 46;
-            byn_l[10] = 1000;
-
-            byn_r[0] = 1000;
-            byn_r[1] = 301;
-            byn_r[2] = 176;
-            byn_r[3] = 125;
-            byn_r[4] = 97;
-            byn_r[5] = 79;
-            byn_r[6] = 67;
-            byn_r[7] = 58;
-            byn_r[8] = 51;
-            byn_r[9] = 46;
-            byn_r[10] = 1000;
-        }
-
-        void initRand()
-        {
-            fpd = 1.0;
-            while (fpd < 16386)
-                fpd = rand() * UINT32_MAX;
-        }
-    };
-
-    inline void Dither::processSamp(double &sample, int chanNum)
-    {
-        if (type == 'n')
-        {
-            njad(sample, chanNum);
-        }
-        else if (type == 't')
-        {
-            tpdf(sample);
-        }
-        else if (type == 'f')
-        {
-            fpdither(sample);
-        }
-    }
-    // Floating point dither for going from double to float
-    // Part of Airwindows plugin suite
-    inline void Dither::fpdither(double &inputSample)
-    {
-        int expon;
-        frexpf((float)inputSample, &expon);
-        fpd ^= fpd << 13;
-        fpd ^= fpd >> 17;
-        fpd ^= fpd << 5;
-        inputSample += (fpd * 3.4e-36l * pow(2, expon + 62)); // removed 'blend' for real use, it's for the demo;
-
-        inputSample = (float)inputSample; // equivalent of 'floor' for 32 bit floating point
-    }
-
-    // Not Just Another Dither
-    // Not truly random. Uses Benford Real Numbers for the dither values
-    // Part of Airwindows plugin suite
-    inline void Dither::njad(double &inputSample, int chanNum)
-    {
-        double *noiseShaping;
-        double(*byn)[13];
-
-        if (chanNum == 0)
-        {
-            noiseShaping = &noise_shaping_l;
-            byn = &byn_l;
-        }
-        else if (chanNum == 1)
-        {
-            noiseShaping = &noise_shaping_r;
-            byn = &byn_r;
-        }
-        else
-        {
-            cerr << "njad only supports a maximum of 2 channels!";
-            throw 1;
-        }
-
-        bool cutbins = false;
-        double drySample = inputSample;
-
-        // Subtract error from previous iteration
-        inputSample -= *noiseShaping;
-
-        // Isolate leading digit of number
-        double benfordize = floor(inputSample);
-        while (benfordize >= 1.0)
-            benfordize /= 10;
-        while (benfordize < 1.0 && benfordize > 0.0000001)
-            benfordize *= 10;
-
-        // Hotbin A becomes the Benford bin value for this number floored (leading digit)
-        int hotbinA = floor(benfordize);
-
-        double totalA = 0;
-        // produce total number- smaller of totalA & totalB is closer to Benford real
-        if ((hotbinA > 0) && (hotbinA < 10))
-        {
-            // Temp add weight to this leading digit
-            *byn[hotbinA] += 1;
-
-            // Coeffs get permanently incremented later in the loop, eventually need
-            // to be scaled back down (cut bins).
-            if (*byn[hotbinA] > 982)
-                cutbins = true;
-
-            totalA += (301 - (*byn[1]));
-            totalA += (176 - (*byn[2]));
-            totalA += (125 - (*byn[3]));
-            totalA += (97 - (*byn[4]));
-            totalA += (79 - (*byn[5]));
-            totalA += (67 - (*byn[6]));
-            totalA += (58 - (*byn[7]));
-            totalA += (51 - (*byn[8]));
-            totalA += (46 - (*byn[9]));
-
-            // Remove temp weight from this leading digit
-            *byn[hotbinA] -= 1;
-        }
-        else
-            hotbinA = 10; // 1000
-
-        // Isolate leading digit of number
-        benfordize = ceil(inputSample);
-
-        while (benfordize >= 1.0)
-            benfordize /= 10;
-        while (benfordize < 1.0 && benfordize > 0.0000001)
-            benfordize *= 10;
-
-        // Hotbin B becomes the Benford bin value for this number ceiled
-        int hotbinB = floor(benfordize);
-
-        double totalB = 0;
-        // produce total number- smaller of totalA & totalB is closer to Benford real
-        if ((hotbinB > 0) && (hotbinB < 10))
-        {
-            // Temp add weight to this leading digit
-            (*byn[hotbinB]) += 1;
-
-            // Coeffs get permanently incremented later in the loop, eventually need
-            // to be scaled back down (cut bins).
-            if (*byn[hotbinB] > 982)
-                cutbins = true;
-
-            totalB += (301 - (*byn[1]));
-            totalB += (176 - (*byn[2]));
-            totalB += (125 - (*byn[3]));
-            totalB += (97 - (*byn[4]));
-            totalB += (79 - (*byn[5]));
-            totalB += (67 - (*byn[6]));
-            totalB += (58 - (*byn[7]));
-            totalB += (51 - (*byn[8]));
-            totalB += (46 - (*byn[9]));
-
-            // Remove temp weight from this leading digit
-            *byn[hotbinB] -= 1;
-        }
-        else
-            hotbinB = 10;
-
-        double outputSample;
-
-        // assign the relevant one to the delay line
-        // and floor/ceil (quantize) signal accordingly
-        if (totalA < totalB)
-        {
-            // Add weight to relevant coeff
-            (*byn[hotbinA]) += 1;
-            outputSample = floor(inputSample);
-        }
-        else
-        {
-            // Add weight to relevant coeff
-            (*byn[hotbinB]) += 1;
-            // If totalB is less, we got here by using a sample that was
-            // previously ceil'd to create hotbin, so add one
-            outputSample = floor(inputSample + 1);
-        }
-
-        if (cutbins)
-        {
-            // Scale down coeffs (weights based on Benford's Law)
-            (*byn[1]) *= 0.99;
-            (*byn[2]) *= 0.99;
-            (*byn[3]) *= 0.99;
-            (*byn[4]) *= 0.99;
-            (*byn[5]) *= 0.99;
-            (*byn[6]) *= 0.99;
-            (*byn[7]) *= 0.99;
-            (*byn[8]) *= 0.99;
-            (*byn[9]) *= 0.99;
-            (*byn[10]) *= 0.99;
-        }
-
-        // Store the error
-        *noiseShaping += outputSample - drySample;
-
-        // Error shouldn't be greater than input sample value
-        if (*noiseShaping > fabs(inputSample))
-            *noiseShaping = fabs(inputSample);
-        if (*noiseShaping < -fabs(inputSample))
-            *noiseShaping = -fabs(inputSample);
-    }
-
-    // TPDF dither
-    inline void Dither::tpdf(double &sample)
-    {
-        double rand1 = ((double)rand()) / ((double)RAND_MAX); // rand value between 0 and 1
-        double rand2 = ((double)rand()) / ((double)RAND_MAX); // rand value between 0 and 1
-        sample += (rand1 - rand2);
-    }
-
     struct ConversionContext
     {
         InputContext inCtx;
@@ -482,8 +67,8 @@ namespace
         inline void scaleAndDither(double &sample, int chanNum);
         inline void writeToBuffer(unsigned char *&out, double &sample, int chanNum);
         template <typename T>
-        inline T clip(T min, T v, T max);
-        static inline int myround(double x);
+        inline T clamp(T min, T v, T max);
+        static inline int myRound(double x);
         inline void writeLSBF(unsigned char *&ptr, unsigned long word);
         static inline void writeFloat(unsigned char *ptr, double sample);
 
@@ -676,7 +261,7 @@ namespace
     };
 
     template <typename T>
-    inline T ConversionContext::clip(T min, T v, T max)
+    inline T ConversionContext::clamp(T min, T v, T max)
     {
         if (v < min)
         {
@@ -699,7 +284,7 @@ namespace
         return v;
     }
 
-    inline int ConversionContext::myround(double x)
+    inline int ConversionContext::myRound(double x)
     {
         // x += x >= 0 ? 0.5 : -0.5;
         return static_cast<int>(round(x));
@@ -722,8 +307,8 @@ namespace
             }
             else
             {
-                writeLSBF(out, clip(-outCtx.peakLevel,
-                                    myround(sample), outCtx.peakLevel - 1));
+                writeLSBF(out, clamp(-outCtx.peakLevel,
+                                    myRound(sample), outCtx.peakLevel - 1));
             }
 
             out += outCtx.channelsNum * outCtx.bytespersample;
@@ -736,7 +321,7 @@ namespace
             }
             else
             {
-                outCtx.pushSamp(clip(-outCtx.peakLevel, myround(sample),
+                outCtx.pushSamp(clamp(-outCtx.peakLevel, myRound(sample),
                                      outCtx.peakLevel - 1),
                                 chanNum);
             }
@@ -884,7 +469,7 @@ int main(int argc, char *argv[])
         try
         {
             inCtx = InputContext(input, format, endianness, inputRate, blockSize,
-                                 channels);
+                                 channels, verboseMode);
         }
         catch (const char *str)
         {
